Initially it would look for the deepest view in a hierarchy which contained current location of the pointer.
There is a pathological case, where you have a pane in front of another pane, where the views of each pane would be in different DOM hierarchies. If two of those views, happened to be visible and contained the region where the pointer was located, then the first view that came out of _dropTargets, would be used to search up in the hierarchy. If it was the wrong hierarchy we will never find the view we were intending to drop on.

This patch assumes that views we are mostly interested in dropping in views in the same pane. So when, sorting the drop targets, first it will sort by pane, and then by hierarchy. This way we will cluster the views that are in the same pane as the source in the beginning positions of the array to be returned.

This fixed this pathological case and has been tested on Chrome, Firefox and Safari.